move code to model and view folders
