+++ sale_exceptions/__init__.py
@@ -23,5 +23,5 @@
 #
 #
 
-from . import sale
+from . import model
 from . import wizard
