+++ sale_exceptions/__openerp__.py
@@ -48,9 +48,10 @@ Contributors
  'author': "Akretion,Odoo Community Association (OCA)",
  'website': 'http://www.akretion.com',
  'depends': ['sale'],
- 'data': ['sale_workflow.xml',
-          'sale_view.xml',
-          'sale_exceptions_data.xml',
+ 'data': ['view/sale_order.xml',
+          'view/sale_exception.xml',
+          'data/workflow.xml',
+          'data/sale_exception.xml',
           'wizard/sale_exception_confirm_view.xml',
           'security/ir.model.access.csv',
           'settings/sale.exception.csv'],
