+++ sale_exceptions/view/sale_order.xml
@@ -2,58 +2,6 @@
 <openerp>
     <data>
 
-        <record id="view_sale_exception_tree" model="ir.ui.view">
-            <field name="name">sale.exception.tree</field>
-            <field name="model">sale.exception</field>
-            <field name="arch" type="xml">
-                <tree string="Sale Exception">
-                    <field name="active"/>
-                    <field name="name"/>
-                    <field name="description"/>
-                    <field name="model"/>
-                    <field name="sequence"/>
-                </tree>
-            </field>
-        </record>
-
-        <record id="view_sale_exception_form" model="ir.ui.view">
-            <field name="name">sale.exception.form</field>
-            <field name="model">sale.exception</field>
-            <field name="arch" type="xml">
-                <form string="Sale Exception Setup">
-                    <group colspan="4" col="2">
-                        <field name="name"/>
-                        <field name="description"/>
-                    </group>
-                    <group col="4" colspan="4" groups="base.group_sale_manager">
-                        <field name="active"/>
-                        <field name="sequence"/>
-                        <group colspan="4" col="2" groups="base.group_system">
-                            <field name="model"/>
-                            <field name="code"/>
-                        </group>
-                    </group>
-                    <group colspan="4" col="2">
-                        <separator string="Affected Sales Orders"/>
-                        <newline/>
-                        <field name="sale_order_ids" nolabel="1" domain="[('state', '=', 'draft')]"/>
-                    </group>
-                </form>
-            </field>
-        </record>
-
-        <record id="action_sale_test_tree" model="ir.actions.act_window">
-                  <field name="name">Exception Rules</field>
-                  <field name="res_model">sale.exception</field>
-                  <field name="view_type">form</field>
-                  <field name="view_mode">tree,form</field>
-                  <field name="view_id" ref="view_sale_exception_tree"/>
-                  <field name="context">{'active_test': False}</field>
-              </record>
-
-        <menuitem action="action_sale_test_tree" id="menu_sale_test" parent="base.menu_sale_config_sales" />
-
-
         <record id="view_order_form" model="ir.ui.view">
             <field name="name">sale_exceptions.view_order_form</field>
             <field name="model">sale.order</field>
@@ -109,5 +57,6 @@
                 </filter>
             </field>
         </record>
+
     </data>
 </openerp>
